added some tests

+++ tests/test_main.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 //extern int test_f(double c);
-extern double func();
+extern int func();
 
 int main(){
     
     //int a = test_f(4);
-    printf("%f\n" , func());
+    printf("%d\n" , func());
     
     return 0;
 }
